perubahan edit task karena perubahan database

+++ src/edittask.php
@@ -32,10 +32,32 @@ mysqli_query($con, "UPDATE tasks
 				WHERE id=$id_task");
 
 // Update assignee
-// Delete all assignees
-mysqli_query($con, "DELETE FROM assignees
-				WHERE task=$id_task");
-// Insert assignees
+// Delete assignee
+$del = mysqli_query($con, "SELECT * 
+						FROM assignees 
+						WHERE task=$id_task");
+if (mysqli_num_rows($del) > 0) {
+	while ($try = mysqli_fetch_array($del)) {
+		// check user id
+		$id_user = $try['member'];
+		$check = mysqli_query($con, "SELECT * 
+								FROM members 
+								WHERE id=$id_user");
+		$user = mysqli_fetch_array($check);
+		$found = false;
+		for ($i = 0; $i < $count_assignee; $i++) {
+			$current = $assignees[$i];
+			if (strcmp($assignees[$i],$user['username']) == 0) {
+				$found = true;
+				break;
+			}
+		}
+		if ($found == false) {
+			mysqli_query($con, "DELETE FROM assignees 
+							WHERE member=$id_user");
+		}
+	}
+}
 for ($i = 0; $i < $count_assignee; $i++) {
 	$current = $assignees[$i];
 	$current = trim($current," ");
@@ -45,9 +67,15 @@ for ($i = 0; $i < $count_assignee; $i++) {
 							WHERE username='$current'");
 	$user = mysqli_fetch_array($check);
 	$id_user = $user['id'];
-	echo $id_user."<br />";
-	mysqli_query($con, "INSERT INTO assignees
-					VALUES ($id_user,$id_task)");
+	// Check user id
+	$check = mysqli_query($con, "SELECT * 
+							FROM assignees 
+							WHERE member=$id_user");
+	$exist = mysqli_num_rows($check);
+	if ($exist == 0) {
+		mysqli_query($con, "INSERT INTO assignees
+						VALUES ($id_user,$id_task,0)");
+	}
 }
 
 // Update tag
